mechanics/requirements.txt
byotrack==1.3.12
contourpy==1.3.3
crc32c==2.7.1.post0
cycler==0.12.1
defusedxml==0.7.1
donfig==0.8.1.post1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
imagecodecs==2025.8.2
imageio==2.37.0
Jinja2==3.1.6
kerchunk==0.2.9
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numcodecs==0.16.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
pylapy==1.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tifffile==2025.10.4
torch==2.9.0
torch-tps==1.2.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
wheel==0.45.1
zarr==3.1.3
jsonargparse